Core rewritten and optimized
