+++ server/api/config/custom-environment-variables.json
@@ -1,3 +1,3 @@
 {
-    "database":"database"
+    "database":"mongodb+srv://stn-admin:aV7JHuEcTtywIE27@cluster0-lvbc6.mongodb.net/stn-database?retryWrites=true&w=majority"
 }
