+++ server/api/config/default.json
@@ -1,4 +1,4 @@
   
 {
-    "jwtPrivateKey":"default"
+    "database":"mongodb+srv://stn-admin:aV7JHuEcTtywIE27@cluster0-lvbc6.mongodb.net/stn-database?retryWrites=true&w=majority"
 }
