+++ server/api/index.js
@@ -6,11 +6,7 @@ const port = process.env.PORT || 3000
 
 //Implementing module of the route
 const aasa = require('./routes/aasa')
-const people = require('./routes/profile') 
-const petition = require('./routes/petition')
-const donation = require('./routes/donation')
-
-
+const main = require('./routes/main') 
 
 var app = express()
 //Disable x-powerd-by header beacuase of security reasons 
@@ -28,16 +24,17 @@ app.get('/', async(req,res) => {
 })
 
 /** Call to  database **/
-mongoose.connect("mongodb+srv://stn-admin:aV7JHuEcTtywIE27@cluster0-lvbc6.mongodb.net/stn-database?retryWrites=true&w=majority")
-.then(() => console.log("Connected to SayTheirName Node Service database \n --------------- \n"))
-.catch(error => console.error("Could not connect to database" + error))
+
+
+mongoose.connect(config.get('database'))
+.then(() => console.log('Connected to SayTheirName Node Service database \n --------------- \n'))
+.catch(error => console.error('Could not connect to database' + error))
 
 
 //Routes
 app.use('/',aasa)
-app.use('/', people)
-app.use('/',petition)
-app.use('/',donation)
+app.use('/', main)
+
 
 app.listen(port, () => console.log(`SayTheirNames-Node Service listening on: ${port}`))
