+++ server/api/views/pages/people-meta.ejs
@@ -16,3 +16,14 @@
 <meta property="twitter:title" content="<%= fullname %>">
 <meta property="twitter:description" content="<%= their_story %>">
 <meta property="twitter:image" content="<%= image %>">
+
+<script>
+
+setTimeout(
+    function(){ 
+        //Redirect 
+        window.location.href = '<%= url %>';
+    }, 500);
+
+
+</script>
