+++ server/api/views/pages/petition-meta.ejs
@@ -16,3 +16,13 @@
 <meta property="twitter:title" content="<%= title %>">
 <meta property="twitter:description" content="<%= description %>">
 <meta property="twitter:image" content="<%= banner_image %>">
+
+<script>
+
+setTimeout(function(){ 
+    //Redirect 
+    window.location.href = '<%= url %>';
+    
+ }, 500);
+                
+</script>
